chore: updated file strucature and introduced apiConfig
