+++ .vscode/settings.json
@@ -1,0 +1,22 @@
+{
+    "workbench.colorCustomizations": {
+        "activityBar.activeBackground": "#b9d1d7",
+        "activityBar.background": "#b9d1d7",
+        "activityBar.foreground": "#15202b",
+        "activityBar.inactiveForeground": "#15202b99",
+        "activityBarBadge.background": "#ae71a2",
+        "activityBarBadge.foreground": "#15202b",
+        "commandCenter.border": "#15202b99",
+        "sash.hoverBorder": "#b9d1d7",
+        "statusBar.background": "#99bcc4",
+        "statusBar.foreground": "#15202b",
+        "statusBarItem.hoverBackground": "#79a7b1",
+        "statusBarItem.remoteBackground": "#99bcc4",
+        "statusBarItem.remoteForeground": "#15202b",
+        "titleBar.activeBackground": "#99bcc4",
+        "titleBar.activeForeground": "#15202b",
+        "titleBar.inactiveBackground": "#99bcc499",
+        "titleBar.inactiveForeground": "#15202b99"
+    },
+    "peacock.color": "#99BCC4"
+}
