+++ src/apiConfig.js
@@ -1,0 +1,12 @@
+
+
+const API_BASE_URL = 'https://q-buddy-backend-1.onrender.com';
+// const API_BASE_URL = 'http://localhost:3000';
+
+export const ENDPOINTS = {
+    SIGNIN: `${API_BASE_URL}/auth/login`,
+    SIGNUP: `${API_BASE_URL}/auth/register`,
+    // Add other endpoints as needed
+};
+
+export default API_BASE_URL;
