+++ src/app/App.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import SignIn from './signIn';
-import SignUp from './signUp';
-import HomePage from './home';
+import SignIn from '../auth/signIn.js';
+import SignUp from '../auth/signUp.js';
+import HomePage from '../pages/home';
 import "leaflet/dist/leaflet.css";
-import DashboardLayoutSlots from './components/DashboardLayoutSlots.tsx';
+import DashboardLayoutSlots from '../components/DashboardLayoutSlots.tsx';
 
 
 function App() {
